Correccion

+++ trunk/Common/StringUtilities.cpp
@@ -105,20 +105,18 @@ namespace stringUtilities {
 		return string.substr(from,to);
 	}
 
-	std::string replaceCharForString(char charR,char* buff,int size,std::string replacer)
-	{
+	std::string replaceCharForString(char charR,char* buff,int size,std::string replacer) {
 		std::string result="";
-		for(int i=0;i<size;i++)
-		{
-			if(buff[i]==charR)
-			{
-				result+=replacer;
+		for (int i=0; i < size; i++) {
+			if(buff[i] == charR) {
+				result += replacer;
 			}
 			else
-				result+=buff[i];
+				result += buff[i];
 		}
 		return result;
 	}
+
 	std::string pairIntToString(std::pair<int, int> key) {
 		std::string aux = intToString(key.first)+","+intToString(key.second);
 		return aux;
@@ -143,34 +141,30 @@ namespace stringUtilities {
 		return auxPair;
 	}
 
-	int replaceStringForChar(char charR,char* buff,std::string replaced,std::string toConvert)
-	{
+	int replaceStringForChar(char charR,char* buff,std::string replaced,std::string toConvert) {
 		std::vector<int> posiciones; 
-		int posicion=toConvert.find(replaced,0);
-		while(posicion!=-1)
+		int posicion = toConvert.find(replaced,0);
+		while(posicion != -1)
 		{
 			posiciones.push_back(posicion);
-			posicion=toConvert.find(replaced,posicion+1);
+			posicion = toConvert.find(replaced,posicion+1);
 		}
-		unsigned i=0;
-		int chars=0;
-		unsigned pos=0;
-		while(i<posiciones.size())
-		{	
-			while(pos<posiciones[i])
-			{
-				buff[chars]=toConvert[pos];
+		unsigned i = 0;
+		int chars = 0;
+		int pos = 0;
+		while( i < posiciones.size()) {	
+			while(pos < posiciones[i]) {
+				buff[chars] = toConvert[pos];
 				chars++;
 				pos++;
 			}
-			buff[chars]=charR;
+			buff[chars] = charR;
 			chars++;
-			pos+=replaced.size();
+			pos += replaced.size();
 			i++;
 		}
-		while(pos<toConvert.size())
-		{
-			buff[chars]=toConvert[pos];
+		while(static_cast<unsigned>(pos) < toConvert.size() ) {
+			buff[chars] = toConvert[pos];
 			chars++;
 			pos++;
 		}
